backend/requirements.txt
# ===== Web 框架 =====
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# ===== 数据库 =====
sqlalchemy==2.0.36
aiomysql==0.2.0
alembic==1.14.0

# ===== 配置管理 =====
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ===== JWT 认证 =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ===== HTTP 客户端（百度 OCR） =====
httpx==0.27.2

# ===== 定时任务 =====
apscheduler==3.10.4

# ===== 测试 =====
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
